Moved backup file name check to backup call

+++ src/Core/Backup/MoveFileBackupStrategy.cs
@@ -2,26 +2,35 @@
 
 namespace Core.Backup;
 
-public class MoveFileBackupStrategy
+public class MoveFileBackupStrategy : IBackupStrategy
 {
-
-    private readonly IFileSystem fs;
-    private readonly Func<string, string> generateBackupFilePath;
-
-    public MoveFileBackupStrategy(IFileSystem fs, Func<string, string> generateBackupFilePath)
+    public interface IBackupFileNaming
     {
-        this.fs = fs;
-        this.generateBackupFilePath = generateBackupFilePath;
+        public string ToBackup(string fullPath);
+        public bool IsBackup(string fullPath);
     }
 
-    public void PerformBackup(string fullPath)
+    private readonly IFileSystem fs;
+    private readonly IBackupFileNaming backupFileNaming;
+
+    public MoveFileBackupStrategy(IFileSystem fs, IBackupFileNaming backupFileNaming)
     {
+        this.fs = fs;
+        this.backupFileNaming = backupFileNaming;
+    }
+
+    public virtual void PerformBackup(string fullPath)
+    {
+        if (backupFileNaming.IsBackup(fullPath))
+        {
+            throw new InvalidOperationException("Installing a backup file is forbidden");
+        }
         if (!fs.File.Exists(fullPath))
         {
             return;
         }
 
-        var backupFilePath = generateBackupFilePath(fullPath);
+        var backupFilePath = backupFileNaming.ToBackup(fullPath);
         if (fs.File.Exists(backupFilePath))
         {
             fs.File.Delete(fullPath);
@@ -34,7 +43,7 @@ public class MoveFileBackupStrategy
 
     public void RestoreBackup(string fullPath)
     {
-        var backupFilePath = generateBackupFilePath(fullPath);
+        var backupFilePath = backupFileNaming.ToBackup(fullPath);
         if (fs.File.Exists(backupFilePath))
         {
             fs.File.Move(backupFilePath, fullPath);
@@ -43,7 +52,7 @@ public class MoveFileBackupStrategy
 
     public void DeleteBackup(string fullPath)
     {
-        var backupFilePath = generateBackupFilePath(fullPath);
+        var backupFilePath = backupFileNaming.ToBackup(fullPath);
         if (fs.File.Exists(backupFilePath))
         {
             fs.File.Delete(backupFilePath);
